[now-cli] Update `now env` to support empty values (#4046)
This PR adds support for empty values when using `now env add|ls|rm|pull`.

This is necessary when using system environment variables such as `NOW_GITHUB_COMMIT_SHA` which will not receive a value from the user but will instead assign a value automatically during deployment.

+++ packages/now-cli/src/util/env/remove-env-record.ts
@@ -10,7 +10,7 @@ export default async function removeEnvRecord(
   target?: ProjectEnvTarget
 ): Promise<void> {
   output.debug(
-    `Removing environment variable ${envName} from target ${target}`
+    `Removing Environment Variable ${envName} from target ${target}`
   );
 
   const qs = target ? `?target=${encodeURIComponent(target)}` : '';
@@ -22,17 +22,17 @@ export default async function removeEnvRecord(
     method: 'DELETE',
   });
 
-  const idOrName = (env.value || '').startsWith('@')
-    ? env.value.slice(1)
-    : env.value;
-  const urlSecret = `/v2/now/secrets/${idOrName}`;
-  const secret = await client.fetch<Secret>(urlSecret);
+  if (env && env.value) {
+    const idOrName = env.value.startsWith('@') ? env.value.slice(1) : env.value;
+    const urlSecret = `/v2/now/secrets/${idOrName}`;
+    const secret = await client.fetch<Secret>(urlSecret);
 
-  // Since integrations add global secrets, we must only delete if the secret was
-  // specifically added to this project
-  if (secret && secret.projectId === projectId) {
-    await client.fetch<Secret>(urlSecret, {
-      method: 'DELETE',
-    });
+    // Since integrations add global secrets, we must only delete if the secret was
+    // specifically added to this project
+    if (secret && secret.projectId === projectId) {
+      await client.fetch<Secret>(urlSecret, {
+        method: 'DELETE',
+      });
+    }
   }
 }
